Create StateOnlyViewModel and EventsOnlyViewModel, make LaunchViewModel a StateOnlyViewModel, tweak ScreenViewModel to hold compose State rather than StateFlow, separate UI state and UI event definitions into different files
