+++ app/src/main/java/io/libzy/ui/connect/ConnectSpotifyScreen.kt
@@ -15,7 +15,6 @@ import androidx.compose.material.Text
 import androidx.compose.material.rememberScaffoldState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
@@ -40,7 +39,7 @@ import kotlinx.coroutines.launch
 @Composable
 fun ConnectSpotifyScreen(navController: NavController, viewModelFactory: ViewModelProvider.Factory) {
     val viewModel: ConnectSpotifyViewModel = viewModel(factory = viewModelFactory)
-    val uiState by viewModel.uiState.collectAsState()
+    val uiState by viewModel.uiState
     val scope = rememberCoroutineScope()
     val scaffoldState = rememberScaffoldState()
