+++ app/src/main/java/io/libzy/ui/connect/ConnectSpotifyUiEvent.kt
@@ -1,0 +1,7 @@
+package io.libzy.ui.connect
+
+enum class ConnectSpotifyUiEvent {
+    SPOTIFY_CONNECTED,
+    SPOTIFY_SCAN_FAILED,
+    SPOTIFY_AUTHORIZATION_FAILED
+}
