+++ app/src/main/java/io/libzy/ui/connect/ConnectSpotifyUiState.kt
@@ -12,9 +12,3 @@ data class ConnectSpotifyUiState(
 ) {
     val libraryScanInProgress = libraryScanState != null && !libraryScanState.isFinished
 }
-
-enum class ConnectSpotifyUiEvent {
-    SPOTIFY_CONNECTED,
-    SPOTIFY_SCAN_FAILED,
-    SPOTIFY_AUTHORIZATION_FAILED
-}
