+++ app/src/main/java/io/libzy/ui/launch/LaunchViewModel.kt
@@ -1,29 +1,36 @@
 package io.libzy.ui.launch
 
 import android.content.Context
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import io.libzy.persistence.prefs.SharedPrefKeys
 import io.libzy.persistence.prefs.getSharedPrefs
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
+import io.libzy.ui.common.StateOnlyViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-// TODO: make this a ScreenViewModel
-class LaunchViewModel @Inject constructor(appContext: Context) : ViewModel() {
+/**
+ * A [StateOnlyViewModel] providing the state for [LaunchScreen] that determines which "home screen" the app
+ * should launch to. This ViewModel produces persistent state rather than instantaneous navigation events so that
+ * the loading time is reduced (collecting events has a higher overhead than reading state), and also because it is
+ * more fitting for the ViewModel's purpose. Once we determine what the home screen should be, that should be considered
+ * a persistent state, not a one-time event, because the navigation is not user-initiated, and will always be the same
+ * if [LaunchScreen] is ever recreated from the back stack, making it useful to already have the correct state set in
+ * the ViewModel rather than having to wait for another navigation event.
+ */
+class LaunchViewModel @Inject constructor(appContext: Context) : StateOnlyViewModel<LaunchUiState>() {
 
-    private val _uiState = MutableStateFlow(LaunchUiState.LOADING)
-    val uiState = _uiState.asStateFlow()
+    override val initialUiState: LaunchUiState = LaunchUiState.LOADING
 
     init {
         viewModelScope.launch {
             val spotifyConnected = appContext.getSharedPrefs().getBoolean(SharedPrefKeys.SPOTIFY_CONNECTED, false)
 
-            _uiState.value = if (spotifyConnected) {
-                LaunchUiState.SPOTIFY_CONNECTED
-            } else {
-                LaunchUiState.NEEDS_SPOTIFY_CONNECTION
+            updateUiState {
+                if (spotifyConnected) {
+                    LaunchUiState.SPOTIFY_CONNECTED
+                } else {
+                    LaunchUiState.NEEDS_SPOTIFY_CONNECTION
+                }
             }
         }
     }
