+++ app/src/main/java/io/libzy/ui/query/QueryUiEvent.kt
@@ -1,0 +1,5 @@
+package io.libzy.ui.query
+
+enum class QueryUiEvent {
+    SUBMIT_QUERY
+}
