+++ app/src/main/java/io/libzy/ui/results/ResultsScreen.kt
@@ -23,7 +23,6 @@ import androidx.compose.material.icons.rounded.StarRate
 import androidx.compose.material.rememberScaffoldState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
@@ -62,7 +61,7 @@ import kotlinx.coroutines.launch
 @Composable
 fun ResultsScreen(navController: NavController, viewModelFactory: ViewModelProvider.Factory, query: Query) {
     val viewModel: ResultsViewModel = viewModel(factory = viewModelFactory)
-    val uiState by viewModel.uiState.collectAsState()
+    val uiState by viewModel.uiState
     val scope = rememberCoroutineScope()
     val scaffoldState = rememberScaffoldState()
