+++ app/src/main/java/io/libzy/ui/results/ResultsUiEvent.kt
@@ -1,0 +1,5 @@
+package io.libzy.ui.results
+
+enum class ResultsUiEvent {
+    SPOTIFY_REMOTE_FAILURE
+}
